+++ examples/builder/logging.rb
@@ -20,6 +20,7 @@ JSONParser =
   .terminator do
     true
   end
+  .hookable
   .build
 
 XMLParser =
@@ -37,6 +38,7 @@ XMLParser =
   .terminator do
     true
   end
+  .hookable
   .build
 
 def build_json_scraper(type)
@@ -50,6 +52,7 @@ def build_json_scraper(type)
         .map { |element| [element['name'], element['emoji']] }
         .to_h
     end
+    .hookable
     .build
 end
 
@@ -69,20 +72,50 @@ def build_xml_scraper(type)
         .map { |element| [element['name'], element['emoji']] }
         .to_h
     end
+    .hookable
     .build
 end
 
 AnimalsXMLScraper = build_xml_scraper(:animals)
 FruitsXMLScraper = build_xml_scraper(:fruits)
 
+loggable = proc do |node|
+  indent_size = 2
+  blank = ' '
+  list_style = '*'
+
+  node.after_matcher! do |matched, _node_:, **|
+    prefix = list_style.rjust(_node_.depth * indent_size - indent_size + list_style.length, blank)
+    puts "#{prefix} #{_node_.identity}: [matched: #{matched}]"
+    matched
+  end
+
+  if node.is_a?(CallableTree::Node::External)
+    input_label  = 'Input :'
+    output_label = 'Output:'
+
+    node
+      .before_caller! do |input, *, _node_:, **|
+        input_prefix = input_label.rjust(_node_.depth * indent_size + input_label.length, blank)
+        puts "#{input_prefix} #{input}"
+        input
+      end
+      .after_caller! do |output, _node_:, **|
+        output_prefix = output_label.rjust(_node_.depth * indent_size + output_label.length, blank)
+        puts "#{output_prefix} #{output}"
+        output
+      end
+  end
+end
+
 tree = CallableTree::Node::Root.new.seekable.append(
-  JSONParser.new.seekable.append(
-    AnimalsJSONScraper.new,
-    FruitsJSONScraper.new
+  JSONParser.new.tap(&loggable).seekable.append(
+    AnimalsJSONScraper.new.tap(&loggable).verbosify,
+    FruitsJSONScraper.new.tap(&loggable).verbosify
   ),
-  XMLParser.new.seekable.append(
-    AnimalsXMLScraper.new,
-    FruitsXMLScraper.new
+  XMLParser.new.tap(&loggable).seekable.append(
+    AnimalsXMLScraper.new.tap(&loggable).verbosify,
+    FruitsXMLScraper.new.tap(&loggable).verbosify
   )
 )