+++ spec/callable_tree/node/builder_spec.rb
@@ -27,17 +27,23 @@ RSpec.describe CallableTree::Node::Builder do
       context 'hookable: true' do
         let(:hookable) { true }
 
-        it { is_expected.to respond_to(:before_call) }
-        it { is_expected.to respond_to(:around_call) }
-        it { is_expected.to respond_to(:after_call) }
+        it { is_expected.to respond_to(%i[before_call before_caller].sample) }
+        it { is_expected.to respond_to(%i[around_call around_caller].sample) }
+        it { is_expected.to respond_to(%i[after_call after_caller].sample) }
+        it { is_expected.to respond_to(:before_matcher) }
+        it { is_expected.to respond_to(:around_matcher) }
+        it { is_expected.to respond_to(:after_matcher) }
       end
 
       context 'hookable: false' do
         let(:hookable) { false }
 
-        it { is_expected.not_to respond_to(:before_call) }
-        it { is_expected.not_to respond_to(:around_call) }
-        it { is_expected.not_to respond_to(:after_call) }
+        it { is_expected.not_to respond_to(%i[before_call before_caller].sample) }
+        it { is_expected.not_to respond_to(%i[around_call around_caller].sample) }
+        it { is_expected.not_to respond_to(%i[after_call after_caller].sample) }
+        it { is_expected.not_to respond_to(:before_matcher) }
+        it { is_expected.not_to respond_to(:around_matcher) }
+        it { is_expected.not_to respond_to(:after_matcher) }
       end
     end
   end